Şuan gecikme çok az
